fix: fe

+++ app/src/main/java/com/example/petshopapplication/API_model/ShipmentSearchResponse.java
@@ -87,10 +87,10 @@ public class ShipmentSearchResponse {
     public static class Parcel {
         private String name;
         private int quantity;
-        private int width;
-        private int height;
-        private int length;
-        private int weight;
+        private double width;
+        private double height;
+        private double length;
+        private double weight;
         private double cweight;
         private String metadata;
